Full game appears to be ported to assemblyscript, we'll see if it works

+++ assembly/engine_v13_wasm/model.ts
@@ -44,7 +44,7 @@ export function createLinearShape(type: string, pattern: string, owner: i32, beg
 }
 
 export class SearchResult {
-  eval: i32 = 0
+  eval: f64 = 0
   evalFlag: string = "exact"  // or "upper-bound" or "lower-bound"
   bestVariation: i32[][] = []
 }
